Added some parts of diffuse material support to CudaTraceAndScatter. Cleaned up the CudaTraceAndScatter and the kernel a bit.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2421 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ maths/mathstypes.h
@@ -588,7 +588,7 @@ inline float fastLog2(float a)
 	};
 	U u;
 	u.f = a;
-	float x = u.i;
+	float x = (float)u.i;
 	//float x = (float)*((int*)&a);
 	x *= 0.00000011920928955078125f; // 2^-23
 	x -= 127.0f;
